Made some new changes after previous fixes by MM.
 * Some of the changes were suggested by AE after MM's previous hotfixes.

 * Some of the changes relate to if RAMIN = 0 or 1 and if -z or -n METDATA.
The changes are to make the resulting scenarios work well.

 * When RAMIN is set to 0 by the user and -n METDATA, then in a few places,
we may get errors due to being unable to delete or make files or folders (e.g.,
if we do not have access to ram), etc. However, these are minor errors and
the code should work overall. These errors were not fixed at the moment, due
to their lower priority.